+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 
+## next version
+
 ## 0.3.0
 
 ### Added